DualValueSetSelect to use ValueSetSelect
