+++ src/types/emr/medicationStatement.ts
@@ -29,13 +29,13 @@ export type MedicationStatement = {
   reason?: string;
 
   medication: Code;
-  dosage_text?: string;
+  dosage_text: string;
   effective_period?: Period;
 
   patient: string; // UUID
   encounter: string; // UUID
 
-  information_source?: MedicationStatementInformationSourceType;
+  information_source: MedicationStatementInformationSourceType;
 
   note?: string;
 };
@@ -44,9 +44,9 @@ export type MedicationStatementRequest = {
   status: MedicationStatementStatus;
   reason?: string;
   medication: Code;
-  dosage_text?: string;
+  dosage_text: string;
   effective_period?: Period;
-  information_source?: MedicationStatementInformationSourceType;
+  information_source: MedicationStatementInformationSourceType;
   note?: string;
 };
 
@@ -55,10 +55,10 @@ export type MedicationStatementRead = {
   status: MedicationStatementStatus;
   reason?: string;
   medication: Code;
-  dosage_text?: string;
+  dosage_text: string;
   effective_period?: Period;
   encounter: string;
-  information_source?: MedicationStatementInformationSourceType;
+  information_source: MedicationStatementInformationSourceType;
   note?: string;
   created_at: string;
   modified_at: string;
